Update Buckling_MF_GP_Separate_Cat.py
Updated all problems to have Matrix encoder and NN encoder versions.
